src/VumbaSoft.AdventureWorks.Controllers/Production/ProductCategories/ProductCategoriesController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class ProductCategoriesController : ValidatedController<IProductCategoryValidator, IProductCategoryService>
    {
        public ProductCategoriesController(IProductCategoryValidator validator, IProductCategoryService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ProductCategoryView category)
        {
            if (!Validator.CanCreate(category))
                return View(category);

            Service.Create(category);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductCategoryView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductCategoryView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ProductCategoryView category)
        {
            if (!Validator.CanEdit(category))
                return View(category);

            Service.Edit(category);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductCategoryView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
